requirements.txt
Django==2.2
djangorestframework==3.9.2
pytest==4.4.0
pytest-cov==2.6.1
codecov
coverage
requests==2.21.0
requests-mock==1.5.2
django-cors-headers
firebase-admin
flake8
mock